backend/ips/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
